parking_management/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
pandas==2.2.3
openpyxl==3.1.3
gunicorn==23.0.0
django-environ==0.12.0